+++ framework/contracts/account/manager/tests/adapters.rs
@@ -38,7 +38,7 @@ fn installing_one_adapter_should_succeed() -> AResult {
     });
 
     // Configuration is correct
-    let adapter_config = staking_adapter.config()?;
+    let adapter_config = staking_adapter.base_config()?;
     assert_that!(adapter_config).is_equal_to(adapter::AdapterConfigResponse {
         ans_host_address: deployment.ans_host.address()?,
         dependencies: vec![],
@@ -364,5 +364,28 @@ fn installing_specific_version_should_install_expected() -> AResult {
     let installed_module: ManagerModuleInfo = modules[1].clone();
     assert_that!(installed_module.id).is_equal_to(adapter1.id());
 
+    Ok(())
+}
+
+#[test]
+fn account_install_adapter() -> AResult {
+    let sender = Addr::unchecked(common::OWNER);
+    let chain = Mock::new(&sender);
+    let deployment = Abstract::deploy_on(chain.clone(), Empty {})?;
+    let account = create_default_account(&deployment.account_factory)?;
+
+    deployment
+        .version_control
+        .claim_namespace(1, "tester".to_owned())?;
+
+    let adapter = BootMockAdapter1V1::new_test(chain);
+    adapter.deploy(V1.parse().unwrap(), MockInitMsg)?;
+    let adapter_addr = account.install_adapter(adapter, &MockInitMsg, None)?;
+    let module_addr = account
+        .manager
+        .module_info(common::mock_modules::adapter_1::MOCK_ADAPTER_ID)?
+        .unwrap()
+        .address;
+    assert_that!(adapter_addr).is_equal_to(module_addr);
     Ok(())
 }