+++ modules/contracts/apps/dca/tests/integration.rs
@@ -7,7 +7,7 @@ use abstract_core::objects::{
 use abstract_core::AbstractError;
 use abstract_core::{app::BaseInstantiateMsg, objects::gov_type::GovernanceDetails};
 use abstract_dca_app::msg::{DCAResponse, Frequency};
-use abstract_dca_app::state::{Config, DCAEntry};
+use abstract_dca_app::state::DCAEntry;
 use abstract_dca_app::{
     contract::{DCA_APP_ID, DCA_APP_VERSION},
     msg::{AppInstantiateMsg, ConfigResponse, InstantiateMsg},
@@ -20,6 +20,7 @@ use abstract_interface::{Abstract, AbstractAccount, AppDeployer, VCExecFns, *};
 use croncat_app::{contract::CRONCAT_ID, AppQueryMsgFns, CroncatApp, CRON_CAT_FACTORY};
 use croncat_integration_testing::test_helpers::set_up_croncat_contracts;
 use croncat_integration_testing::DENOM;
+use cw_asset::AssetInfo;
 // Use prelude to get all the necessary imports
 use cw_orch::{anyhow, deploy::Deploy, prelude::*};
 
@@ -72,6 +73,13 @@ fn setup() -> anyhow::Result<(
 
     // Deploy Abstract to the mock
     let abstr_deployment = Abstract::deploy_on(mock.clone(), sender.to_string())?;
+    abstr_deployment.ans_host.execute(
+        &abstract_core::ans_host::ExecuteMsg::UpdateAssetAddresses {
+            to_add: vec![("denom".to_owned(), AssetInfo::native(DENOM).into())],
+            to_remove: vec![],
+        },
+        None,
+    )?;
     // Deploy wyndex to the mock
     let wyndex = wyndex_bundle::WynDex::deploy_on(mock.clone(), Empty {})?;
     // Deploy dex adapter to the mock
@@ -143,7 +151,7 @@ fn setup() -> anyhow::Result<(
                 ans_host_address: abstr_deployment.ans_host.addr_str()?,
             },
             module: AppInstantiateMsg {
-                native_denom: DENOM.to_owned(),
+                native_asset: AssetEntry::new("denom"),
                 dca_creation_amount: Uint128::new(5_000_000),
                 refill_threshold: Uint128::new(1_000_000),
                 max_spread: Decimal::percent(30),
@@ -199,12 +207,10 @@ fn successful_install() -> anyhow::Result<()> {
     assert_eq!(
         config,
         ConfigResponse {
-            config: Config {
-                native_denom: DENOM.to_owned(),
-                dca_creation_amount: Uint128::new(5_000_000),
-                refill_threshold: Uint128::new(1_000_000),
-                max_spread: Decimal::percent(30),
-            }
+            native_asset: AssetEntry::from("denom"),
+            dca_creation_amount: Uint128::new(5_000_000),
+            refill_threshold: Uint128::new(1_000_000),
+            max_spread: Decimal::percent(30),
         }
     );
     Ok(())
@@ -315,7 +321,7 @@ fn create_dca_convert_negative() -> anyhow::Result<()> {
             offer_asset = OfferAsset::new(USD, 100_u128),
             target_asset = AssetEntry::new(USD),
             e = AbstractError::from(StdError::generic_err(
-                "asset pairing wynd/usd,usd not found in ans_host"
+                "asset pairing wyndex/usd,usd not found in ans_host"
             ))
         )),
     );
@@ -477,7 +483,7 @@ fn update_dca_negative() -> anyhow::Result<()> {
             offer_asset = OfferAsset::new(USD, 200_u128),
             target_asset = AssetEntry::new(USD),
             e = AbstractError::from(StdError::generic_err(
-                "asset pairing wynd/usd,usd not found in ans_host"
+                "asset pairing wyndex/usd,usd not found in ans_host"
             ))
         )),
     );